requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pyquery==2.0.0
pandas==2.1.0
openpyxl==3.1.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
httpx==0.25.2 